# openApiSchema.yml
openapi: 3.0.0
info:
  title: 股票数据管理API
  description: |
    提供股票数据管理、用户认证和用户数据管理功能的API接口。
    包括股票数据的CRUD操作、用户注册登录、股票选择等功能。
  version: '3.0.0'
  contact:
    name: Stockstock
    email: support@example.com
servers:
  - url: http://localhost:3000
    description: 本地开发环境
  - url: https://api.example.com
    description: 生产环境
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 使用JWT令牌进行认证，格式为 "token"
  schemas:
...
    Prediction:
      type: object
      properties:
        stockId:
          type: string
          description: 股票代码
        date:
          type: string
          description: 预测日期
        predictedPrice:
          type: number
          description: 预测价格
        createdAt:
          type: string
          format: date-time
          description: 预测创建时间
      required:
        - stockId
        - date
        - predictedPrice
        - createdAt
        
    # 添加因子预测相关模式
    FactorPrediction:
      type: object
      properties:
        id:
          type: integer
          description: 预测记录ID
        stockId:
          type: string
          description: 股票代码
        date:
          type: string
          description: 预测日期
        TTM:
          type: number
          description: 预测的市盈率TTM值
        PE:
          type: number
          description: 预测的市盈率值
        PB:
          type: number
          description: 预测的市净率值
        PCF:
          type: number
          description: 预测的市现率值
        created_at:
          type: string
          format: date-time
          description: 预测创建时间
      required:
        - stockId
        - date
        - TTM
        - PE
        - PB
        - PCF
    ModelPredictionResponse:
      type: object
      properties:
        message:
          type: string
          description: 操作结果消息
        pricePredictions:
          type: object
          properties:
            count:
              type: integer
              description: 价格预测结果总数
            samples:
              type: array
              items:
                $ref: '#/components/schemas/Prediction'
              description: 样例预测结果(最多5条)
        factorPredictions:
          type: object
          properties:
            count:
              type: integer
              description: 因子预测结果总数
            samples:
              type: array
              items:
                $ref: '#/components/schemas/FactorPrediction'
              description: 样例因子预测结果(最多5条)
      required:
        - message
        - pricePredictions
        - factorPredictions
    
    # 通用响应模式
    Error:
      type: object
      properties:
        error:
          type: string
          description: 错误消息
        code:
          type: string
          description: 错误代码
...
paths:
...
  /modelPredict:
    post:
      tags:
        - 模型预测
      summary: 执行模型预测
      description: |
        训练模型并执行预测，存储：
        1. 股票估值预测 - 预测未来估值
        2. 预测中间变量 - 预测未来的TTM、PE、PB、PCF等因子值
        
        预测结果会保存到数据库中。如果数据库中已存在相同股票相同日期的预测，会更新而非追加记录。
        此操作需要管理员权限。
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 预测成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelPredictionResponse'
        '401':
          description: 未提供认证或认证已失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足，需要管理员权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 未找到股票数据
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /modelPredict/{stockId}:
    get:
      tags:
        - 模型预测
      summary: 获取股票预测历史
      description: 获取指定股票的所有预测历史记录
      security:
        - bearerAuth: []
      parameters:
        - name: stockId
          in: path
          required: true
          description: 股票代码
          schema:
            type: string
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Prediction'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 未找到股票代码的预测数据
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /modelPredict/factor/{stockId}:
    get:
      tags:
        - 模型预测
      summary: 获取股票因子预测历史
      description: 获取指定股票的所有因子预测历史记录，包括TTM、PE、PB、PCF等指标
      security:
        - bearerAuth: []
      parameters:
        - name: stockId
          in: path
          required: true
          description: 股票代码
          schema:
            type: string
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FactorPrediction'
        '401':
          description: 未提供认证或认证已失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 未找到预测结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'